Create Conefor imputs for connectivity analysis
R code for creating conefor imputs for spatial analyses. This code is an easy way for creating the imputs for Conefor analyses for each shape file. It help me a lot when creating imputs in GIS software such as ArcGis Pro or QGIS files for conefor were build with some errors.